collect UTC offset at startup, before starting any other threads
Yes DST happening while the server is running can cause the offset to be wrong. That's why we show the offset. An if we really care, we should always be able to restart the server. And if this somehow became a persistent issue we could have a DST switch in the UI. I don't expect that to happen though.